requirements.txt
Django==5.0.2
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
django-jazzmin==3.0.1
django-unfold==0.57.0
djangorestframework==3.16.0
Pillow==11.2.1
reportlab==4.4.0
xhtml2pdf==0.2.17
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.0
pyyaml==6.0.2
certifi==2025.4.26
chardet==5.2.0
idna==3.10
urllib3==2.4.0
charset-normalizer==3.4.2
html5lib==1.1
arabic-reshaper==3.0.0
python-bidi==0.6.6
webencodings==0.5.1
pypdf==5.4.0
